Move failure detail table to panel folder
