+++ static/app/views/starfish/views/webServiceView/failureDetailPanel/failureDetailTable.tsx
